Propose a new architecture for algorithms (#403)
* Propose a new architecture for algorithms

* Remove working file

* Improve consistency of names
