+++ src/leaspy/algo/algo_with_annealing.py
@@ -4,14 +4,13 @@ import numpy as np
 
 from leaspy.exceptions import LeaspyAlgoInputError
 
-from ..settings import AlgorithmSettings
+from .settings import AlgorithmSettings
 
-__all__ = ["AlgoWithAnnealingMixin"]
+__all__ = ["AlgorithmWithAnnealingMixin"]
 
 
-class AlgoWithAnnealingMixin:
-    """
-    Mixin class to use in algorithms that requires `temperature_inv`.
+class AlgorithmWithAnnealingMixin:
+    """Mixin class to use in algorithms that requires `temperature_inv`.
 
     Note that this mixin should be used with a class inheriting from `AbstractAlgo`, which must have `algo_parameters`
     attribute.
