--- src/leaspy/algo/others/__init__.py
@@ -1,9 +1,0 @@
-from .constant_prediction_algo import ConstantPredictionAlgorithm
-from .lme_fit import LMEFitAlgorithm
-from .lme_personalize import LMEPersonalizeAlgorithm
-
-__all__ = [
-    "ConstantPredictionAlgorithm",
-    "LMEFitAlgorithm",
-    "LMEPersonalizeAlgorithm",
-]
