+++ src/leaspy/algo/personalize/mean_posterior.py
@@ -5,15 +5,13 @@ import torch
 from leaspy.utils.typing import DictParamsTorch
 
 from ..base import AlgorithmName
-from .abstract_mcmc_personalize import AbstractMCMCPersonalizeAlgo
+from .mcmc import McmcPersonalizeAlgorithm
 
-__all__ = ["MeanPosterior"]
+__all__ = ["MeanPosteriorAlgorithm"]
 
 
-class MeanPosterior(AbstractMCMCPersonalizeAlgo):
-    """
-    Sampler-based algorithm that derives individual parameters as the most frequent mean posterior value from `n_iter` samplings.
-
+class MeanPosteriorAlgorithm(McmcPersonalizeAlgorithm):
+    """Sampler-based algorithm that derives individual parameters as the most frequent mean posterior value from `n_iter` samplings.
 
     Parameters
     ----------
