+++ src/leaspy/algo/personalize/mode_posterior.py
@@ -5,14 +5,13 @@ import torch
 from leaspy.utils.typing import DictParamsTorch
 
 from ..base import AlgorithmName
-from .abstract_mcmc_personalize import AbstractMCMCPersonalizeAlgo
+from .mcmc import McmcPersonalizeAlgorithm
 
-__all__ = ["ModePosterior"]
+__all__ = ["ModePosteriorAlgorithm"]
 
 
-class ModePosterior(AbstractMCMCPersonalizeAlgo):
-    """
-    Sampler-based algorithm that derives individual parameters as the most frequent mode posterior value from `n_iter` samplings.
+class ModePosteriorAlgorithm(McmcPersonalizeAlgorithm):
+    """Sampler-based algorithm that derives individual parameters as the most frequent mode posterior value from `n_iter` samplings.
 
     TODO? we could derive some confidence intervals on individual parameters thanks to this personalization algorithm...
