--- src/leaspy/algo/utils/__init__.py
@@ -1,9 +1,0 @@
-from .algo_with_annealing import AlgoWithAnnealingMixin
-from .algo_with_device import AlgoWithDeviceMixin
-from .algo_with_samplers import AlgoWithSamplersMixin
-
-__all__ = [
-    "AlgoWithAnnealingMixin",
-    "AlgoWithDeviceMixin",
-    "AlgoWithSamplersMixin",
-]
